Return Integer values in estimated_minutes_watched
Since YouTube never returns decimal digits anyways.

+++ lib/yt/models/channel.rb
@@ -112,7 +112,7 @@ module Yt
       has_report :uniques, Integer
 
       # @macro report_by_channel_dimensions
-      has_report :estimated_minutes_watched, Float
+      has_report :estimated_minutes_watched, Integer
 
       # @macro report_by_gender_and_age_group
       has_report :viewer_percentage, Float
