+++ lib/yt/models/playlist.rb
@@ -157,7 +157,7 @@ module Yt
       has_report :views, Integer
 
       # @macro report_by_playlist_dimensions
-      has_report :estimated_minutes_watched, Float
+      has_report :estimated_minutes_watched, Integer
 
       # @macro report_by_gender_and_age_group
       has_report :viewer_percentage, Float
